LayoutTests/fast/dom/HTMLElement/attr-dir-auto-change-text.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
<script src="../../js/resources/js-test-pre.js"></script>
<style>
.testDiv {-webkit-border-start: 5px solid green; -webkit-border-end: 5px solid red; }
</style>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="result"></div>

<div id="test" dir="auto" class="testDiv">Test test test</div>
<script>

description('Test that directionality is re-evaluated when an element with dir=auto set, has the text of its first child changed.');

var el = document.getElementById("test");
shouldBe("document.defaultView.getComputedStyle(el, null).getPropertyValue('border-right-color')", "'rgb(255, 0, 0)'");
el.innerHTML = "מקור השם עברית";
shouldBe("document.defaultView.getComputedStyle(el, null).getPropertyValue('border-right-color')", "'rgb(0, 128, 0)'");
el.style.display = "none";
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
