third_party/WebKit/LayoutTests/fast/dom/HTMLElement/attr-dir-auto-change-text-expected.txt

Test that directionality is re-evaluated when an element with dir=auto set, has the text of its first child changed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.defaultView.getComputedStyle(el, null).getPropertyValue('border-right-color') is 'rgb(255, 0, 0)'
PASS document.defaultView.getComputedStyle(el, null).getPropertyValue('border-right-color') is 'rgb(0, 128, 0)'
PASS successfullyParsed is true

TEST COMPLETE
